Refactor discovery controller - use the aws-app-mesh-controller-for-k8s types - split App Mesh operations from controller logic - remove Kubernetes service controller
